Added cors middleware func to run in api routes

+++ pages/api/track/[id].ts
@@ -1,9 +1,11 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import prisma from '@/hooks/prisma'
+import { runMiddleware } from '@/middleware/corsMiddleware'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     const trackId = Number(req?.query?.id)
     if (req.method === 'PATCH') {
+        await runMiddleware(req, res)
         let track = await prisma?.powerHourSong?.update({
             where: {
                 id: trackId
@@ -12,6 +14,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
         })
         res.status(200).json({ track })
     } else if (req.method === 'DELETE') {
+        await runMiddleware(req, res)
         let track = await prisma?.powerHourSong?.findUnique({
             where: {
                 id: trackId
@@ -48,7 +51,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
                 }
             })
             res.status(200).json({ updatedPowerHour })
-        } else {
+        } else if (req.method === 'DELETE') {
+            await runMiddleware(req, res)
             await prisma.powerHourSong.delete({
                 where: {
                     id: trackId
